indicator: volatility-based indicator that identify stop loss exit points for long and short trading positions
Chandelier Exit is based on the Average True Range (ATR) indicator. It is designed to keep traders in the trend until a defined trend reversal happens. Traders use CE to maximize their returns in a trade and make stop loss exit decisions. It is based on the principle that a trend reversal is highly probable when the price of an asset moves against an existing trend up to three times the average volatility. It uses the high and low prices over a defined period of time to compute the CE value.